requirements.txt
prometheus_client
requests~=2.28.2
numpy~=1.24.2
matrix-nio
Pillow~=9.4.0
python-magic~=0.4.27
numpy
beautifulsoup4~=4.11.2
flask~=2.2.3
icinga2api~=0.6.1
urllib3~=1.26.14
aiofiles~=0.6.0
markdown